Move the `encodeToXmlString` helper function to `src/core/core_utils.js`
With the previous patch this function is now *only* accessed on the worker-thread, hence it's no longer necessary to include it in the *built* `pdf.js` file.
