+++ src/shared/util.js
@@ -967,53 +967,6 @@ const createObjectURL = (function createObjectURLClosure() {
   };
 })();
 
-const XMLEntities = {
-  /* < */ 0x3c: "&lt;",
-  /* > */ 0x3e: "&gt;",
-  /* & */ 0x26: "&amp;",
-  /* " */ 0x22: "&quot;",
-  /* ' */ 0x27: "&apos;",
-};
-
-function encodeToXmlString(str) {
-  const buffer = [];
-  let start = 0;
-  for (let i = 0, ii = str.length; i < ii; i++) {
-    const char = str.codePointAt(i);
-    if (0x20 <= char && char <= 0x7e) {
-      // ascii
-      const entity = XMLEntities[char];
-      if (entity) {
-        if (start < i) {
-          buffer.push(str.substring(start, i));
-        }
-        buffer.push(entity);
-        start = i + 1;
-      }
-    } else {
-      if (start < i) {
-        buffer.push(str.substring(start, i));
-      }
-      buffer.push(`&#x${char.toString(16).toUpperCase()};`);
-      if (char > 0xd7ff && (char < 0xe000 || char > 0xfffd)) {
-        // char is represented by two u16
-        i++;
-      }
-      start = i + 1;
-    }
-  }
-
-  if (buffer.length === 0) {
-    return str;
-  }
-
-  if (start < str.length) {
-    buffer.push(str.substring(start, str.length));
-  }
-
-  return buffer.join("");
-}
-
 export {
   AbortException,
   AnnotationActionEventType,
@@ -1035,7 +988,6 @@ export {
   createPromiseCapability,
   createValidAbsoluteUrl,
   DocumentActionEventType,
-  encodeToXmlString,
   escapeString,
   FONT_IDENTITY_MATRIX,
   FontType,
